requirements.txt
discord.py
configparser
